No more days modules
Removes the `days/mod.ts`, as the functions are no longer needed in 2
separate files and can be consolidated into the `days/main.ts` script
instead.
